% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aftsur.R
\name{aftsur}
\alias{aftsur}
\title{Semi-parametric Accelerated Failure Time Mixture Cure Model}
\usage{
aftsur(
  formula_aft,
  formula_logit,
  cens_status,
  y_L,
  y_R,
  data,
  lambda = 1e-05,
  knots = NULL
)
}
\arguments{
\item{formula_aft}{a formula for AFT model ~ X1 + X2.}

\item{formula_logit}{a formula for AFT model ~ Z1 + Z2.}

\item{cens_status}{a formula for censored status in the data with sequence: event, left, right, interval censoring. ~ delta + deltaL + deltaR + deltaI.}

\item{y_L}{a vector of left censored times.}

\item{y_R}{a vector of right censored times.}

\item{data}{a data frame which includes survival times and variables for the AFT model and the logistic regression.}

\item{lambda}{initial value for smoothing parameter. By default, lambda = 1e-5.}

\item{knots}{the number of Gaussian basis function. If leave it as NULL, knots will be automatically chosen based on the sample size.}
}
\value{
fit parameter estimates, standard deviations and
}
\description{
A method to fit semi-parametric AFT mixture cure model using the maximum penalised likelihood estimation, where left, right, interval-censored data are allowed.
}
