% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.aftsur.R
\name{print.summary.aftsur}
\alias{print.summary.aftsur}
\title{Printing Summary Output}
\usage{
\method{print}{summary.aftsur}(x)
}
\arguments{
\item{x}{an object of class "summary.aftsur".}
}
\description{
A function that prints the summary output for a model result.
}
