% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ptces}
\alias{ptces}
\title{Simulated Partly interval-censored Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 100 rows and 12 columns.
}
\usage{
data(ptces)
}
\description{
Description_of_the_Data
}
\keyword{datasets}
