% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.aftsur.R
\name{summary.aftsur}
\alias{summary.aftsur}
\title{Summarising a Model Fit for Semi-parametric Accelerated Failure Time Mixture Cure Model}
\usage{
\method{summary}{aftsur}(fit)
}
\arguments{
\item{fit}{an object of class "aftsur", a model fitted by the \code{aftsur} function.}
}
\description{
\code{summary} method for class \code{"aftsur"}.
}
\examples{
require(survival)
# load data
data("simu_data")
# create a formula using a Surv
formula_aft <- Surv(y_L, y_R, type = "interval2") ~ X1 + X2 - 1
# fit a model
aft_fit <- aftsur(formula = formula_aft, cure_var = ~ Z1 + Z2 + Z3, offset = TRUE, data = simu_data)
summary(aft_fit)
}
