% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aftsur.R
\name{aftsur}
\alias{aftsur}
\title{Semi-parametric Accelerated Failure Time Mixture Cure Model}
\usage{
aftsur(formula, cure_var, offset = FALSE, lambda = 1e-05, knots = NULL, data)
}
\arguments{
\item{formula}{a formula object to specify the covariates for the latency part of the mixture cure model. The response variable must be a \code{Surv} object obtained from the \code{survival} package.}

\item{cure_var}{a formula object in which to specify the covariates for the incidence part of the mixture cure model.}

\item{offset}{if offset is \code{FALSE}, an intercept term will be added into the incidence covariates. By default, \code{offset = FALSE}.}

\item{lambda}{an initial value for the smoothing parameter. By default, \code{lambda = 1e-5}.}

\item{knots}{an integer which specifies the number of basis functions used to estimate the baseline hazard function. By default, \code{knots = 4} if the sample size is less than 500; \code{knots = 5} if the sample size is greater than 500 but less than 1000 and \code{knots = 6} if the sample size is greater or equal to 1000.}

\item{data}{a data frame which includes survival times, covariates, censoring statuses. Covairates can be numerical or categorical. If categorical variables are supplied, these variables should be a class of \code{Factor}. Censoring statuses need to be either \code{0} or \code{1}.}
}
\value{
\code{aftsur} returns an object of class \code{"aftsur"}.
}
\description{
A method to fit  a semi-parametric accelerated failure time (AFT) mixture cure model using maximum penalised likelihood estimation. This method allows for the inclusion of left, right, and interval-censored data. Logistic regression is employed to model the incidence.
}
\examples{
require(survival)
# load data
data("simu_data")
# create a formula using a Surv
formula_aft <- Surv(y_L, y_R, type = "interval2") ~ X1 + X2 - 1
# fit a model
aftsur(formula = formula_aft, cure_var = ~ Z1 + Z2 + Z3, offset = TRUE, data = simu_data)
}
