% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aftsur.R
\name{aftsur}
\alias{aftsur}
\title{Semi-parametric Accelerated Failure Time Mixture Cure Model}
\usage{
aftsur(formula, cure_var, offset = FALSE, lambda = 1e-05, knots = NULL, data)
}
\arguments{
\item{formula}{a formula object which specifies the covariates for the latency part of the mixture cure model. The response must be a \code{Surv} object which is from \code{survival} package.}

\item{cure_var}{a formula object which specifies the covariates for the incidence part of the mixture cure model.}

\item{offset}{if offset is \code{FALSE}, an intercept term will be added into the incidence covariates. By default, \code{offset = FALSE}.}

\item{lambda}{an initial value for the smoothing parameter. By default, lambda = $10^{-5}$.}

\item{knots}{an integer which specifies the number of basis functions used to estimate the baseline hazard function. By default, \code{knots = 4} if the sample size is less than 500; \code{knots = 5} if the sample size is greater than 500 but less than 1000 and \code{knots = 6} if the sample size is greater or equal to 1000.}

\item{data}{a data frame which includes survival times, covariates, censoring status.}
}
\value{
\code{aftsur} returns an object of class \code{"aftsur"}.
}
\description{
A method to fit a semi-parametric AFT mixture cure model using maximum penalised likelihood estimation, where left, right and interval-censored data are allowed. Logistic regression is used to model the incidence.
}
\examples{
require(survival)
# load data
data("ptces")
# create Surv object
formula_aft <- Surv(y_L, y_R, type = "interval2") ~ X1 + X2 + X3 - 1
# fit model
aftsur(formula = formula_aft, cure_var = ~ Z1 + Z2 + Z3, offset = TRUE, data = ptces)
}
