% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aftsur.R
\name{aftsur}
\alias{aftsur}
\title{Semi-parametric Accelerated Failure Time Mixture Cure Model}
\usage{
aftsur(formula, cure_var, offset = FALSE, lambda = 1e-05, knots = NULL, data)
}
\arguments{
\item{formula}{a formula with latency covariates on the right hand side of "~" and a \code{Surv} object as the response.}

\item{cure_var}{a formula specifies incidence covariates.}

\item{offset}{if offset is \code{FALSE}, an intercept term will be added into the incidence covariates. By default, \code{offset = FALSE}.}

\item{lambda}{initial value for smoothing parameter. By default, \code{lambda} = 1e-5.}

\item{knots}{the number of Gaussian basis functions. If leave it as NULL, knots will be automatically chosen based on the sample size.}

\item{data}{a data frame which includes survival times, covariates, censoring status.}
}
\value{
\code{aftsur} returns an object of class \code{"aftsur"}.
}
\description{
A method to fit semi-parametric AFT mixture cure model using the maximum penalised likelihood estimation, where left, right, interval-censored data are allowed. A logistic regression is used to model the incidence.
}
\examples{
require(survival)
# load data
data("ptces")
# create Surv object
formula_aft <- Surv(y_L, y_R, type = "interval2") ~ X1 + X2 + X3 - 1
# fit model
aftsur(formula = formula_aft, cure_var = ~ Z1 + Z2 + Z3, offset = TRUE, data = ptces)
}
