% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hz.R
\name{plot_hz}
\alias{plot_hz}
\title{Plot Predicted Hazard Function}
\usage{
plot_hz(fit, x = NULL)
}
\arguments{
\item{fit}{an object of class "aftsur". A model fitted by the \code{aftsur} function.}

\item{x}{an one-row dataframe which contains latency variables with the same data structure as the training data. If \code{x = NULL}, the baseline hazard will be plotted.}
}
\value{
a \code{ggplot2} object
}
\description{
a function that plots predicted hazard rates from the semi-parametric AFT mixture cure model.
}
\examples{
require(survival)
# load data
data("simu_data")
# create a formula using a Surv
formula_aft <- Surv(y_L, y_R, type = "interval2") ~ X1 + X2 - 1
# fit a model
aft_fit <- aftsur(formula = formula_aft, cure_var = ~ Z1 + Z2 + Z3, offset = TRUE, data = simu_data)
plot_hz(aft_fit)
}
