% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hz.R
\name{plot_hz}
\alias{plot_hz}
\title{Plot Predicted Hazard Function}
\usage{
plot_hz(fit, x = NULL)
}
\arguments{
\item{fit}{a model fitted by the \code{aftsur} function.}

\item{x}{a vector of latency covariates that is used to estimate hazard rates. If \code{x = NULL}, the baseline hazard will be predicted.}
}
\description{
a function that plots predicted hazard rates from the semi-parametric AFT mixture cure model.
}
