% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sur.R
\name{plot_sur}
\alias{plot_sur}
\title{Predicted Survival Function Plot}
\usage{
plot_sur(fit, x = NULL)
}
\arguments{
\item{fit}{a model fitted by the \code{aftsur} function.}

\item{x}{a vector of latency covariates that is used to estimate survival probabilities. If \code{x = NULL}, the baseline survival will be estimated.}
}
\description{
plot for predicted survival probabilities from the semi-parametric AFT mixture cure model.
}
