% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.aftsur.R
\name{print.summary.aftsur}
\alias{print.summary.aftsur}
\title{Summarising a Model Fit for Semi-parametric Accelerated Failure Time Mixture Cure Model}
\usage{
\method{print}{summary.aftsur}(x)
}
\arguments{
\item{x}{an object of class "summary.aftsur".}
}
\description{
a function that summaries model result.
}
