% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ptces}
\alias{ptces}
\title{Simulated Partly Interval-censored Data}
\format{
This dataset includes 100 observations.
\describe{
\item{\code{X1 X2 X3}}{incidence covariates}
\item{\code{Z1 Z2 Z3}}{latency covariates, where \code{Z1} is an intercept term}
\item{\code{delta deltaL deltaL deltaR}}{indicators for event, left, right, and interval-censoring}
\item{\code{y_L y_R}}{event and censoring times. For events, \code{y_L = y_R}. If the observation is left-censored, \code{y_L = NA}. If the observation is right-censored, \code{y_R = NA}.}
}
}
\usage{
data(ptces)
}
\description{
A simulated partly interval-censored dataset.
}
\keyword{datasets}
